Fix compilation warnings
git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/kdesupport/taglib@1019655 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ taglib/mpeg/id3v2/frames/attachedpictureframe.cpp
@@ -137,7 +137,7 @@ void AttachedPictureFrame::parseFields(const ByteVector &data)
 
   d->mimeType = readStringField(data, String::Latin1, &pos);
   /* Now we need at least two more bytes available */	
-  if (pos + 1 >= data.size()) {
+  if (uint(pos) + 1 >= data.size()) {
     debug("Truncated picture frame.");
     return;
   }
